+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 - Avoid error in `creating_group_split` when a group is not correct.
   [sgeulette]
+- Taken into account routing table in `IncomingEmail` consumer.
+  [sgeulette]
 
 0.20.0 (2023-11-28)
 -------------------